Branch HU1 new release.
Se modificó la rama HU1 para implementar las recomendaciones dadas por los mentores: Jhoan, Emily, Matteo Velásquez y Juan Frasser.

1. Ahora los método de crear categoría tienen su lógica en el CategoryUseCase.java y no en el CategoryJpaAdapter.java como lo tenías antes.

2. Se pasa a tener las validaciones simples como campos vacíos o nulos y temas de tamaño dentro de CategoryDto.java y se quitan de CategoryUseCase.java para eliminar exceso de código inecesario.

3. Estos nuevos cambios pasan a ser validados en los Test

4. Como se movieron validaciones del CategoryJpaAdapter.java a la clase CategoryUseCase.java. Ahora dentro de los Test las validaciones de Guardado pasan de estar en CategoryJpaAdapterTest.java a estar en CategoryUseCaseTest.java
